+++ src/app/common/tooltips/compacting.text.ts
@@ -1,4 +1,4 @@
-export class CompactingTooltipsTexts {
+export class CompactionTooltipsTexts {
     public static readonly MAX_SLICE_SIZE = "Maximum size of a single data slice file.";
     public static readonly MAX_SLICE_RECORDS = "Maximum amount of records in a single data slice file";
 }