+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/compacting/dataset-settings-compacting-tab.component.spec.ts
@@ -17,7 +17,7 @@ import {
 } from "src/app/common/base-test.helpers.spec";
 import { mockDatasetBasicsDerivedFragment, mockDatasetBasicsRootFragment } from "src/app/search/mock.data";
 import { ModalService } from "src/app/components/modal/modal.service";
-import { DatasetCompactingService } from "../../services/dataset-compacting.service";
+import { DatasetCompactionService } from "../../services/dataset-compaction.service";
 import { of } from "rxjs";
 import { NavigationService } from "src/app/services/navigation.service";
 import AppValues from "src/app/common/app.values";
@@ -27,7 +27,7 @@ describe("DatasetSettingsCompactingTabComponent", () => {
     let component: DatasetSettingsCompactingTabComponent;
     let fixture: ComponentFixture<DatasetSettingsCompactingTabComponent>;
     let modalService: ModalService;
-    let datasetCompactingService: DatasetCompactingService;
+    let datasetCompactionService: DatasetCompactionService;
     let navigationService: NavigationService;
 
     beforeEach(async () => {
@@ -72,7 +72,7 @@ describe("DatasetSettingsCompactingTabComponent", () => {
         fixture = TestBed.createComponent(DatasetSettingsCompactingTabComponent);
         component = fixture.componentInstance;
         modalService = TestBed.inject(ModalService);
-        datasetCompactingService = TestBed.inject(DatasetCompactingService);
+        datasetCompactionService = TestBed.inject(DatasetCompactionService);
         navigationService = TestBed.inject(NavigationService);
         component.datasetBasics = mockDatasetBasicsRootFragment;
         fixture.detectChanges();
@@ -96,15 +96,15 @@ describe("DatasetSettingsCompactingTabComponent", () => {
 
     it("should check run hard compacting", fakeAsync(() => {
         const navigateToDatasetViewSpy = spyOn(navigationService, "navigateToDatasetView");
-        const runHardCompactingSpy = spyOn(datasetCompactingService, "runHardCompacting").and.returnValue(of(true));
+        const runHardCompactionSpy = spyOn(datasetCompactionService, "runHardCompaction").and.returnValue(of(true));
         const modalServiceSpy = spyOn(modalService, "error").and.callFake((options) => {
             options.handler?.call(undefined, true);
             return Promise.resolve("");
         });
-        emitClickOnElementByDataTestId(fixture, "run-compacting-btn");
+        emitClickOnElementByDataTestId(fixture, "run-compaction-btn");
 
         expect(modalServiceSpy).toHaveBeenCalledTimes(1);
-        expect(runHardCompactingSpy).toHaveBeenCalledTimes(1);
+        expect(runHardCompactionSpy).toHaveBeenCalledTimes(1);
         tick(AppValues.SIMULATION_START_CONDITION_DELAY_MS);
         expect(navigateToDatasetViewSpy).toHaveBeenCalledTimes(1);
         flush();