+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/compacting/dataset-settings-compacting-tab.component.ts
@@ -2,12 +2,12 @@ import { NavigationService } from "./../../../../../services/navigation.service"
 import { ChangeDetectionStrategy, Component, Input, OnInit } from "@angular/core";
 import { AbstractControl, FormBuilder, Validators } from "@angular/forms";
 import { RxwebValidators } from "@rxweb/reactive-form-validators";
-import { CompactingConditionInput, DatasetBasicsFragment, DatasetFlowType } from "src/app/api/kamu.graphql.interface";
+import { CompactionConditionInput, DatasetBasicsFragment, DatasetFlowType } from "src/app/api/kamu.graphql.interface";
 import { logError, promiseWithCatch } from "src/app/common/app.helpers";
-import { CompactingTooltipsTexts } from "src/app/common/tooltips/compacting.text";
+import { CompactionTooltipsTexts } from "src/app/common/tooltips/compacting.text";
 import { ModalService } from "src/app/components/modal/modal.service";
-import { CompactingMode, SliceUnit, sliceSizeMapper } from "./dataset-settings-compacting-tab.types";
-import { DatasetCompactingService } from "../../services/dataset-compacting.service";
+import { CompactionMode, SliceUnit, sliceSizeMapper } from "./dataset-settings-compacting-tab.types";
+import { DatasetCompactionService } from "../../services/dataset-compaction.service";
 import { DatasetViewTypeEnum } from "src/app/dataset-view/dataset-view.interface";
 import AppValues from "src/app/common/app.values";
 import { BaseComponent } from "src/app/common/base.component";
@@ -20,80 +20,79 @@ import { BaseComponent } from "src/app/common/base.component";
 })
 export class DatasetSettingsCompactingTabComponent extends BaseComponent implements OnInit {
     @Input() public datasetBasics: DatasetBasicsFragment;
-    public hardCompactingForm = this.fb.group({
-        mode: [CompactingMode.FULL, [Validators.required]],
+    public hardCompactionForm = this.fb.group({
+        mode: [CompactionMode.FULL, [Validators.required]],
         sliceUnit: [SliceUnit.MB, [Validators.required]],
         sliceSize: [300, [Validators.required, RxwebValidators.minNumber({ value: 1 })]],
         recordsCount: [10000, [Validators.required, RxwebValidators.minNumber({ value: 1 })]],
         recursive: [{ value: true, disabled: true }],
     });
     public readonly SliceUnit: typeof SliceUnit = SliceUnit;
-    public readonly MAX_SLICE_SIZE_TOOLTIP = CompactingTooltipsTexts.MAX_SLICE_SIZE;
-    public readonly MAX_SLICE_RECORDS_TOOLTIP = CompactingTooltipsTexts.MAX_SLICE_RECORDS;
+    public readonly MAX_SLICE_SIZE_TOOLTIP = CompactionTooltipsTexts.MAX_SLICE_SIZE;
+    public readonly MAX_SLICE_RECORDS_TOOLTIP = CompactionTooltipsTexts.MAX_SLICE_RECORDS;
     public readonly MIN_VALUE_ERROR_TEXT = "The value must be positive";
-    public readonly CompactingMode: typeof CompactingMode = CompactingMode;
-    public compactingArgs: CompactingConditionInput;
+    public readonly CompactionMode: typeof CompactionMode = CompactionMode;
 
     constructor(
         public modalService: ModalService,
         private fb: FormBuilder,
-        private datasetCompactingService: DatasetCompactingService,
+        private datasetCompactionService: DatasetCompactionService,
         private navigationService: NavigationService,
     ) {
         super();
     }
 
     ngOnInit(): void {
-        this.hardCompactingModeChanges();
+        this.hardCompactionModeChanges();
     }
 
-    public get hardCompactingMode(): AbstractControl {
-        return this.hardCompactingForm.controls.mode;
+    public get hardCompactionMode(): AbstractControl {
+        return this.hardCompactionForm.controls.mode;
     }
 
     public get recursive(): AbstractControl {
-        return this.hardCompactingForm.controls.recursive;
+        return this.hardCompactionForm.controls.recursive;
     }
 
     public get sliceUnit(): AbstractControl {
-        return this.hardCompactingForm.controls.sliceUnit;
+        return this.hardCompactionForm.controls.sliceUnit;
     }
 
     public get recordsCount(): AbstractControl {
-        return this.hardCompactingForm.controls.recordsCount;
+        return this.hardCompactionForm.controls.recordsCount;
     }
 
     public get sliceSize(): AbstractControl {
-        return this.hardCompactingForm.controls.sliceSize;
+        return this.hardCompactionForm.controls.sliceSize;
     }
 
     public get sliceSizeInBytes(): number {
         return (
-            (this.hardCompactingForm.controls.sliceSize.value as number) *
-            sliceSizeMapper[this.hardCompactingForm.controls.sliceUnit.value as SliceUnit]
+            (this.hardCompactionForm.controls.sliceSize.value as number) *
+            sliceSizeMapper[this.hardCompactionForm.controls.sliceUnit.value as SliceUnit]
         );
     }
 
     public get sliceSizeControl(): AbstractControl {
-        return this.hardCompactingForm.controls.sliceSize;
+        return this.hardCompactionForm.controls.sliceSize;
     }
 
     public get recordsCountControl(): AbstractControl {
-        return this.hardCompactingForm.controls.recordsCount;
+        return this.hardCompactionForm.controls.recordsCount;
     }
 
-    private hardCompactingModeChanges(): void {
+    private hardCompactionModeChanges(): void {
         this.trackSubscription(
-            this.hardCompactingMode.valueChanges.subscribe((value: CompactingMode) => {
+            this.hardCompactionMode.valueChanges.subscribe((value: CompactionMode) => {
                 switch (value) {
-                    case CompactingMode.FULL: {
+                    case CompactionMode.FULL: {
                         this.disableAndClearControl(this.recursive);
                         this.sliceUnit.enable();
                         this.recordsCount.enable();
                         this.sliceSize.enable();
                         break;
                     }
-                    case CompactingMode.METADATA_ONLY: {
+                    case CompactionMode.METADATA_ONLY: {
                         this.recursive.enable();
                         this.disableAndClearControl(this.sliceUnit);
                         this.disableAndClearControl(this.recordsCount);
@@ -101,29 +100,29 @@ export class DatasetSettingsCompactingTabComponent extends BaseComponent impleme
                         break;
                     }
                     default: {
-                        logError("Unknown CompactingMode key");
+                        logError("Unknown CompactionMode key");
                     }
                 }
             }),
         );
     }
 
-    public onRunCompacting(): void {
+    public onRunCompaction(): void {
         promiseWithCatch(
             this.modalService.error({
-                title: "Run compacting",
-                message: "Do you want to run hard compacting?",
+                title: "Run compaction",
+                message: "Do you want to run hard compaction?",
                 yesButtonText: "Ok",
                 noButtonText: "Cancel",
                 handler: (ok) => {
                     if (ok) {
                         this.trackSubscription(
-                            this.datasetCompactingService
-                                .runHardCompacting({
+                            this.datasetCompactionService
+                                .runHardCompaction({
                                     datasetId: this.datasetBasics.id,
-                                    datasetFlowType: DatasetFlowType.HardCompacting,
-                                    compactingArgs: this.setCompactingArgs(
-                                        this.hardCompactingMode.value as CompactingMode,
+                                    datasetFlowType: DatasetFlowType.HardCompaction,
+                                    compactionArgs: this.setCompactionArgs(
+                                        this.hardCompactionMode.value as CompactionMode,
                                     ),
                                 })
                                 .subscribe((result: boolean) => {
@@ -150,16 +149,16 @@ export class DatasetSettingsCompactingTabComponent extends BaseComponent impleme
         control.markAsPristine();
     }
 
-    private setCompactingArgs(mode: CompactingMode): CompactingConditionInput {
+    private setCompactionArgs(mode: CompactionMode): CompactionConditionInput {
         switch (mode) {
-            case CompactingMode.FULL:
+            case CompactionMode.FULL:
                 return {
                     full: {
                         maxSliceSize: this.sliceSizeInBytes,
-                        maxSliceRecords: this.hardCompactingForm.controls.recordsCount.value as number,
+                        maxSliceRecords: this.hardCompactionForm.controls.recordsCount.value as number,
                     },
                 };
-            case CompactingMode.METADATA_ONLY:
+            case CompactionMode.METADATA_ONLY:
                 return {
                     metadataOnly: {
                         recursive: this.recursive.value as boolean,